+++ resources/views/group/item-component.blade.php
@@ -1,0 +1,5 @@
+@livewire(
+    $this->itemComponentClassName(),
+    $this->itemsParams[$itemId] ?? [],
+    key($itemId)
+)